+++ CHANGELOG.md
@@ -21,3 +21,6 @@
 
 ## 0.0.7
 - Added test cases
+- Optimized the `okto.openBottomSheet()`
+- Fixed rendering issue in the okto bottm sheet
+- Added an optional height parameter for the bottom sheet.